# Exchange loop order to optimize executation time of CPU version