--- a/gemm_cpu_naive.cc
+++ b/gemm_cpu_naive.cc
@@ -20,8 +20,8 @@ void gemm_cpu_naive(
 	const int m,
 	const int k
 ) {
-	for (int l = 0; l < k; ++l)
-		for (int i = 0; i < n; ++i)
+	for (int i = 0; i < n; ++i)
+		for (int l = 0; l < k; ++l)
 			for (int j = 0; j < m; ++j) {
 				C[i * m + j] += A[i * k + l] * B[l * m + j];
 			}