The change is about locality.

In my machine,
```
before:
       cpu_native          195.00 us
       cpu_simd            76.39 us
after:
       cpu_naive           193.00 us
       cpu_simd            64.62 us
```

Certainly this is not important at all because the course focuses on GPU programming.